// src/lib/scripts/types.ts
export interface InterfaceKisAnimateConstructorData {
	items?: string | NodeListOf<HTMLElement>;
	type?: string;
	timingFunction?: string;
	delay?: number;
	delayStep?: number;
	duration?: number;
	offset?: string;
	threshold?: number;
}

export interface InterfaceKisAnimateTextConstructorData {
	items?: string | NodeListOf<HTMLElement>;
	type?: string;
	timingFunction?: string;
	delay?: number;
	symbolDelay?: number;
	duration?: number;
	offset?: string;
	threshold?: number;
}